# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Registo e acompanhamento de encomendas de produtos numa empresa"
  version: "2.0.0"
  title: "Gestor de Encomendas Api"
host: "localhost:4000"
basePath: "/gestorencomendas-api/"
schemes:
- "http"
- "https"
paths:
  /utilizador:
    get:
      tags:
      - "Utilizador"
      summary: "Mostrar todos os utilizadores"
      description: "Mostrar todos os utilizadores"
      operationId: "mostrarUtilizador"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Utilizadores mostrados!"
          schema:
            type: "array"
            items:
              type: "string"
        "404":
          description: "Os utilizadores não foram encontrados!"
      x-swagger-router-controller: "Utilizador"
    post:
      tags:
      - "Utilizador"
      summary: "Registar utilizador"
      description: "Registar um novo utilizador"
      operationId: "registarUtilizador"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "utilizador_"
        required: true
        schema:
          $ref: "#/definitions/utilizador_"
      responses:
        "201":
          description: "Registo efetuado com sucesso!"
        "204":
          description: "Erro sem conteúdo!"
        "226":
          description: "Erro em uso!"
        "400":
          description: "Requisição inválida!"
      x-swagger-router-controller: "Utilizador"
  /utilizador/{username}:
    get:
      tags:
      - "Utilizador"
      summary: "Buscar utilizador"
      description: "Buscar um utilizador"
      operationId: "buscarUtilizadorPeloNome"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "Username do utilizador a buscar"
        required: true
        type: "string"
      responses:
        "201":
          description: "Buscado com sucesso!"
          schema:
            $ref: "#/definitions/utilizador_"
        "400":
          description: "Requisição inválida!"
        "404":
          description: "Utilizador não encontrado"
        "500":
          description: "Erro interno no servidor!"
      x-swagger-router-controller: "Utilizador"
    put:
      tags:
      - "Utilizador"
      summary: "Atualizar utilizador"
      description: "Atualizar um utilizador"
      operationId: "atualizarUtilizador"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "Username do utilizador a ser atualizado"
        required: true
        type: "string"
      - in: "body"
        name: "utilizador_"
        description: "Objeto utilizador atualizado"
        required: true
        schema:
          $ref: "#/definitions/utilizador_"
      responses:
        "200":
          description: "Atualizado com sucesso!"
        "400":
          description: "Requisição inválida!"
        "404":
          description: "Utilizador não encontrado"
        "500":
          description: "Erro interno no servidor!"
      x-swagger-router-controller: "Utilizador"
  /utilizador/login:
    post:
      tags:
      - "Utilizador"
      summary: "Faz o login na API"
      description: "Faz o login na API"
      operationId: "loginUtilizador"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "login_"
        description: "Username do utilizador para o login"
        required: true
        schema:
          $ref: "#/definitions/login_"
      responses:
        "200":
          description: "Login efetuado com sucesso!"
          schema:
            type: "string"
        "400":
          description: "Username e/ou password inválido"
      x-swagger-router-controller: "Utilizador"
  /utilizador/logout:
    get:
      tags:
      - "Utilizador"
      summary: "Faz o logout da API"
      description: "Faz o logout da API"
      operationId: "logoutUtilizador"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        default:
          description: "Logout realizado com sucesso"
      x-swagger-router-controller: "Utilizador"
  /utilizador/apagar/{username}:
    delete:
      tags:
      - "Utilizador"
      summary: "Apagar utilizador"
      description: "Apagar um utilizador"
      operationId: "apagarUtilizador"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "O nome do utilizador a ser eliminado"
        required: true
        type: "string"
      responses:
        "201":
          description: "Utilizador apagado com sucesso!"
        "204":
          description: "Erro sem conteúdo!"
        "226":
          description: "Erro em uso!"
        "400":
          description: "Nome inválido!"
        "404":
          description: "Utilizador não encontrado!"
        "500":
          description: "Erro interno no servidor!"
      x-swagger-router-controller: "Utilizador"
  /produto:
    get:
      tags:
      - "Produto"
      summary: "Mostrar todos os produtos"
      description: "Mostrar todos os produtos"
      operationId: "mostrarProduto"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Produtos mostrados!"
          schema:
            type: "array"
            items:
              type: "string"
        "404":
          description: "Os produtos não foram encontrados!"
      x-swagger-router-controller: "Produto"
    post:
      tags:
      - "Produto"
      summary: "Adicionar produto"
      description: "Adicionar um novo produto"
      operationId: "adicionarProduto"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "produto_"
        description: "Produto a ser adicionado à loja"
        required: true
        schema:
          $ref: "#/definitions/produto_"
      responses:
        "201":
          description: "Produto adicionado com sucesso!"
        "204":
          description: "Erro sem conteúdo!"
        "226":
          description: "Erro em uso!"
        "400":
          description: "Requisição inválida!"
        "500":
          description: "Erro interno no servidor!"
      x-swagger-router-controller: "Produto"
  /produto/{id_produto}:
    put:
      tags:
      - "Produto"
      summary: "Atualizar um produto"
      description: "Atualizar um produto"
      operationId: "atualizarProduto"
      produces:
      - "application/json"
      parameters:
      - name: "id_produto"
        in: "path"
        description: "ID do produto a ser atualizado"
        required: true
        type: "string"
      - in: "body"
        name: "produto_"
        description: "Objeto produto a ser atualizado"
        required: true
        schema:
          $ref: "#/definitions/produto_"
      responses:
        "200":
          description: "Atualizado com sucesso!"
        "400":
          description: "Requisição inválida!"
        "404":
          description: "Produto não encontrado"
      x-swagger-router-controller: "Produto"
    delete:
      tags:
      - "Produto"
      summary: "Apaga produto pelo ID"
      description: "Apaga produto pelo ID"
      operationId: "apagarProdutoId"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id_produto"
        in: "path"
        description: "ID do produto a ser apagado"
        required: true
        type: "string"
      responses:
        "200":
          description: "Produto apagado com sucesso!"
        "400":
          description: "ID do produto inválido!"
        "404":
          description: "Produto não encontrado!"
      x-swagger-router-controller: "Produto"
  /encomenda:
    get:
      tags:
      - "Encomenda"
      summary: "Mostrar todas as encomendas"
      description: "Mostrar todas as encomendas"
      operationId: "mostrarEncomenda"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Encomendas mostrados!"
          schema:
            type: "array"
            items:
              type: "string"
        "404":
          description: "As encomendas não foram encontradas!"
      x-swagger-router-controller: "Encomenda"
    post:
      tags:
      - "Encomenda"
      summary: "Registar encomenda"
      description: "Registar uma nova encomenda"
      operationId: "registarEncomenda"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "encomenda_"
        description: "Objeto encomenda"
        required: true
        schema:
          $ref: "#/definitions/encomenda_"
      responses:
        "200":
          description: "Encomenda realizada com sucesso!"
        "400":
          description: "Encomenda inválida!"
      x-swagger-router-controller: "Encomenda"
  /encomenda/{id_encomenda}:
    get:
      tags:
      - "Encomenda"
      summary: "Encontrar encomenda pelo ID"
      description: "Encontrar encomenda pelo ID"
      operationId: "encontrarEncomenda"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id_encomenda"
        in: "path"
        description: "ID da encomenda a encontrar"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Encomenda encontrada com sucesso!"
          schema:
            $ref: "#/definitions/encomenda_"
        "400":
          description: "ID da encomenda invalido!"
        "404":
          description: "Encomenda não encontrada!"
      x-swagger-router-controller: "Encomenda"
    patch:
      tags:
      - "Encomenda"
      summary: "Atualizar uma encomenda pelo id"
      description: "Atualizar uma encomenda"
      operationId: "atualizarEncomenda"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id_encomenda"
        in: "path"
        description: "Objeto encomenda a ser atualizada"
        required: true
        type: "string"
      - in: "body"
        name: "atualizarEncomenda_"
        description: "Information to update the order"
        required: false
        schema:
          $ref: "#/definitions/atualizarEncomenda_"
      responses:
        "201":
          description: "Atualizado com sucesso!"
        "400":
          description: "Requisição inválida!"
        "404":
          description: "Encomenda não encontrada!"
      x-swagger-router-controller: "Encomenda"
definitions:
  login_:
    type: "string"
    properties:
      username:
        type: "string"
      password:
        type: "string"
  utilizador_:
    type: "object"
    properties:
      id_funcionario:
        type: "string"
      funcionario:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      username:
        type: "string"
      datanasc:
        type: "string"
        format: "date"
      cargo:
        type: "string"
        enum:
        - "funcionário"
        - "administrador"
    example:
      password: "password"
      datanasc: "2000-01-23"
      id_funcionario: "id_funcionario"
      funcionario: "funcionario"
      cargo: "funcionário"
      email: "email"
      username: "username"
  produto_:
    type: "object"
    properties:
      id_produto:
        type: "string"
      nome:
        type: "string"
      tipo:
        type: "string"
      preço:
        type: "number"
        format: "double"
      descrição:
        type: "string"
      disp:
        type: "string"
        enum:
        - "Disponível"
        - "Esgotado"
      stock:
        type: "integer"
        format: "int32"
  encomenda_:
    type: "object"
    properties:
      id_encomenda:
        type: "integer"
      estado:
        type: "string"
        enum:
        - "Criada"
        - "Enviada"
        - "Cancelada"
        - "Entregue"
      data_cria:
        type: "string"
        format: "date"
    example:
      estado: "Criada"
      data_cria: "2000-01-23"
      id_encomenda: 0
  atualizarEncomenda_:
    type: "object"
    required:
    - "estado"
    properties:
      estado:
        type: "string"
        enum:
        - "Criada"
        - "Enviada"
        - "Cancelada"
        - "Entregue"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"